requirements.txt
appnope==0.1.3
asgiref==3.7.2
astroid==3.0.0
asttokens==2.2.1
backcall==0.2.0
bleach==6.1.0
blinker==1.7.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.36.0
cryptography==41.0.5
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.7
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.6
Django==3.2
django-allauth==0.58.2
django-cloudinary-storage==0.3.0
django-crispy-forms==1.14.0
django-summernote==0.8.20.0
djangorestframework==3.14.0
executing==1.2.0
Flask==3.0.0
flask-validators==1.0
gunicorn==21.2.0
idna==3.4
ipython==8.12.0
isort==5.12.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
langid==1.1.6
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mccabe==0.7.0
numpy==1.26.2
oauthlib==3.2.2
packaging==23.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.11.0
prompt-toolkit==3.0.38
psycopg2==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pygame==2.5.0
Pygments==2.15.0
PyJWT==2.8.0
pylint==3.0.1
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
SQLAlchemy==2.0.23
sqlparse==0.4.4
stack-data==0.6.2
tomlkit==0.12.1
traitlets==5.9.0
typing_extensions==4.9.0
urllib3==2.1.0
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==3.0.1
